move math to types/math
