+++ x/ecocredit/server/query_server.go
@@ -3,8 +3,8 @@ package server
 import (
 	"github.com/regen-network/regen-ledger/types"
 
-	"github.com/regen-network/regen-ledger/math"
 	"github.com/regen-network/regen-ledger/orm"
+	"github.com/regen-network/regen-ledger/types/math"
 	"github.com/regen-network/regen-ledger/x/ecocredit"
 )
